+++ SampleImplementation/CombatAction.cs
