+++ SampleImplementation/CombatResult.cs
