+++ SampleImplementation/EnemyGroup.cs
