+++ SampleImplementation/GameEngine.cs
