+++ SampleImplementation/GameEngine2.cs
