+++ SampleImplementation/GlobalDropTable.cs
