+++ SampleImplementation/Tests/GameEngine2Tests/GlobalCounter.cs
@@ -2,7 +2,7 @@
 using FluentAssertions.Execution;
 using Xunit;
 
-namespace ZeldaKata.GameEngine2Tests;
+namespace ZeldaKata.Tests.GameEngine2Tests;
 
 public class Global
 {
