+++ SampleImplementation/ZeldaKata.sln
