+++ SampleImplementation/ZeldaKata.sln.DotSettings
